refactor: 더미 서비스 및 더미 응답 리팩토링
